{"types":"C","runs":9}

+++ problem21.java
@@ -3,7 +3,7 @@ import acm.program.*;
 
 public class problem21 extends GraphicsProgram{
 	public void run(){
-		double h = getHeight();
+		int h = getHeight();
 		for(int i = 0; i < 11; i++){
 		GLine hLine = new GLine(0, i * h / 10, 0, i * h / 10);
 		add(hLine);
